requirements.txt
app==0.0.1
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-rest-auth==0.9.5
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
idna==2.8
jsonify==0.5
mysqlclient==1.4.2.post1
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
setuptools==40.9.0
